requirements.txt.txt
#GUI
streamlit

#Gemini and LlammaIndex
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-gemini
